requirements.txt
supabase
resend
python-dotenv